Move user classes into package
default access variables that only other user classes should touch
