+++ src/us/kbase/auth2/lib/user/LocalUser.java
@@ -1,4 +1,4 @@
-package us.kbase.auth2.lib;
+package us.kbase.auth2.lib.user;
 
 import static us.kbase.auth2.lib.Utils.nonNull;
 
@@ -8,6 +8,13 @@ import java.util.Collections;
 import java.util.Set;
 
 import com.google.common.base.Optional;
+
+import us.kbase.auth2.lib.DisplayName;
+import us.kbase.auth2.lib.EmailAddress;
+import us.kbase.auth2.lib.PolicyID;
+import us.kbase.auth2.lib.Role;
+import us.kbase.auth2.lib.UserDisabledState;
+import us.kbase.auth2.lib.UserName;
 
 /** A local user.
  *
