+++ src/us/kbase/test/auth2/lib/storage/mongo/MongoStoragePasswordTest.java
@@ -15,8 +15,6 @@ import org.junit.Test;
 import com.google.common.base.Optional;
 
 import us.kbase.auth2.lib.DisplayName;
-import us.kbase.auth2.lib.LocalUser;
-import us.kbase.auth2.lib.NewUser;
 import us.kbase.auth2.lib.UserName;
 import us.kbase.auth2.lib.exceptions.NoSuchLocalUserException;
 import us.kbase.auth2.lib.exceptions.NoSuchUserException;
@@ -24,6 +22,8 @@ import us.kbase.auth2.lib.identity.RemoteIdentityDetails;
 import us.kbase.auth2.lib.identity.RemoteIdentityID;
 import us.kbase.auth2.lib.identity.RemoteIdentityWithLocalID;
 import us.kbase.auth2.lib.storage.mongo.Fields;
+import us.kbase.auth2.lib.user.LocalUser;
+import us.kbase.auth2.lib.user.NewUser;
 import us.kbase.test.auth2.TestCommon;
 
 public class MongoStoragePasswordTest extends MongoStorageTester {
