+++ src/us/kbase/test/auth2/lib/storage/mongo/MongoStorageRolesTest.java
@@ -14,13 +14,13 @@ import java.util.UUID;
 import org.junit.Test;
 
 import us.kbase.auth2.lib.DisplayName;
-import us.kbase.auth2.lib.NewUser;
 import us.kbase.auth2.lib.Role;
 import us.kbase.auth2.lib.UserName;
 import us.kbase.auth2.lib.exceptions.NoSuchUserException;
 import us.kbase.auth2.lib.identity.RemoteIdentityDetails;
 import us.kbase.auth2.lib.identity.RemoteIdentityID;
 import us.kbase.auth2.lib.identity.RemoteIdentityWithLocalID;
+import us.kbase.auth2.lib.user.NewUser;
 import us.kbase.test.auth2.TestCommon;
 
 public class MongoStorageRolesTest extends MongoStorageTester {
